// web/src/generated/PostMessage.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: PostMessage
// ====================================================

export interface PostMessage {
  postMessage: number;
}

export interface PostMessageVariables {
  user: string;
  content: string;
}